refactor: 스플래시 스크린을 메인에서 로그인 액티비티로 이동하여 적용
