+++ android/app/src/main/java/com/yagubogu/presentation/login/LoginActivity.kt
@@ -2,22 +2,34 @@ package com.yagubogu.presentation.login
 
 import android.content.Intent
 import android.os.Bundle
+import android.util.Log
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import androidx.lifecycle.lifecycleScope
 import com.yagubogu.databinding.ActivityLoginBinding
 import com.yagubogu.presentation.MainActivity
+import kotlinx.coroutines.launch
 
 class LoginActivity : AppCompatActivity() {
     private val binding: ActivityLoginBinding by lazy {
         ActivityLoginBinding.inflate(layoutInflater)
     }
+    private var isAppInitialized = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        setupSplash()
         super.onCreate(savedInstanceState)
         setupView()
         setupListeners()
+        performInitialization(savedInstanceState)
+    }
+
+    private fun setupSplash() {
+        val splashScreen = installSplashScreen()
+        splashScreen.setKeepOnScreenCondition { !isAppInitialized }
     }
 
     private fun setupView() {
@@ -34,6 +46,20 @@ class LoginActivity : AppCompatActivity() {
         binding.constraintBtnGoogle.setOnClickListener {
             startActivity(Intent(this, MainActivity::class.java))
             finish()
+        }
+    }
+
+    private fun performInitialization(savedInstanceState: Bundle?) {
+        lifecycleScope.launch {
+            try {
+                // Todo : 초기화 작업 수행, (LoginViewModel에서 초기 Api 요청, 데이터베이스 조회 등)
+            } catch (e: Exception) {
+                Log.e("LoginActivity", "초기화 실패", e)
+            } finally {
+                isAppInitialized = true
+                if (savedInstanceState == null) {
+                }
+            }
         }
     }
 }
